Solver helper functions in a separate file
