+++ include/samurai/petsc/linear_solver.hpp
@@ -527,38 +527,38 @@ namespace samurai
          * Helper functions
          */
 
-        template <class Scheme, std::enable_if_t<Scheme::cfg_t::scheme_type != SchemeType::NonLinear, bool> = true>
-        auto make_solver(const Scheme& scheme)
-        {
-            return SingleFieldSolver<Assembly<Scheme>>(scheme);
-        }
+        // template <class Scheme, std::enable_if_t<Scheme::cfg_t::scheme_type != SchemeType::NonLinear, bool> = true>
+        // auto make_solver(const Scheme& scheme)
+        // {
+        //     return SingleFieldSolver<Assembly<Scheme>>(scheme);
+        // }
 
-        template <class Scheme>
-        void solve(const Scheme& scheme, typename Scheme::field_t& unknown, typename Scheme::field_t& rhs)
-        {
-            auto solver = make_solver(scheme);
-            solver.solve(unknown, rhs);
-        }
+        // template <class Scheme>
+        // void solve(const Scheme& scheme, typename Scheme::field_t& unknown, typename Scheme::field_t& rhs)
+        // {
+        //     auto solver = make_solver(scheme);
+        //     solver.solve(unknown, rhs);
+        // }
 
-        template <class Scheme, class E>
-        void solve(const Scheme& scheme, typename Scheme::field_t& unknown, const field_expression<E>& rhs_expression)
-        {
-            typename Scheme::field_t rhs = rhs_expression;
-            solve(scheme, unknown, rhs);
-        }
+        // template <class Scheme, class E>
+        // void solve(const Scheme& scheme, typename Scheme::field_t& unknown, const field_expression<E>& rhs_expression)
+        // {
+        //     typename Scheme::field_t rhs = rhs_expression;
+        //     solve(scheme, unknown, rhs);
+        // }
 
-        template <bool monolithic, std::size_t rows, std::size_t cols, class... Operators>
-        auto make_solver(const BlockOperator<rows, cols, Operators...>& block_operator)
-        {
-            return BlockSolver<monolithic, rows, cols, Operators...>(block_operator);
-        }
+        // template <bool monolithic, std::size_t rows, std::size_t cols, class... Operators>
+        // auto make_solver(const BlockOperator<rows, cols, Operators...>& block_operator)
+        // {
+        //     return BlockSolver<monolithic, rows, cols, Operators...>(block_operator);
+        // }
 
-        template <std::size_t rows, std::size_t cols, class... Operators>
-        auto make_solver(const BlockOperator<rows, cols, Operators...>& block_operator)
-        {
-            static constexpr bool default_monolithic = true;
-            return make_solver<default_monolithic, rows, cols, Operators...>(block_operator);
-        }
+        // template <std::size_t rows, std::size_t cols, class... Operators>
+        // auto make_solver(const BlockOperator<rows, cols, Operators...>& block_operator)
+        // {
+        //     static constexpr bool default_monolithic = true;
+        //     return make_solver<default_monolithic, rows, cols, Operators...>(block_operator);
+        // }
 
     } // end namespace petsc
 } // end namespace samurai
