+++ include/samurai/petsc/nonlinear_solver.hpp
@@ -383,11 +383,11 @@ namespace samurai
          * Helper functions
          */
 
-        template <class Scheme, std::enable_if_t<Scheme::cfg_t::scheme_type == SchemeType::NonLinear, bool> = true>
-        auto make_solver(const Scheme& scheme)
-        {
-            return SingleFieldNonLinearSolver<Assembly<Scheme>>(scheme);
-        }
+        // template <class Scheme, std::enable_if_t<Scheme::cfg_t::scheme_type == SchemeType::NonLinear, bool> = true>
+        // auto make_solver(const Scheme& scheme)
+        // {
+        //     return SingleFieldNonLinearSolver<Assembly<Scheme>>(scheme);
+        // }
 
     } // end namespace petsc
 } // end namespace samurai
